revert to int

+++ src/n2knn.cpp
@@ -29,7 +29,7 @@ using std::chrono::high_resolution_clock;
 typedef Eigen::Triplet<double> T;
 
 // [[Rcpp::export]]
-Eigen::SparseMatrix<double> n2Knn(const Rcpp::NumericMatrix& m, int64_t k, int64_t nThreads=10, bool verbose=true, std::string indexType="angular", int64_t M=12, int64_t MaxM0=24, float ef_search_multiplier=50, bool quiet=false) {
+Eigen::SparseMatrix<double> n2Knn(const Rcpp::NumericMatrix& m, int k, int nThreads=10, bool verbose=true, std::string indexType="angular", int M=12, int MaxM0=24, float ef_search_multiplier=50, bool quiet=false) {
   Eigen::SparseMatrix<double> mat(m.nrow(), m.nrow());
 
   if(m.nrow() <= k) {
@@ -43,7 +43,7 @@ Eigen::SparseMatrix<double> n2Knn(const Rcpp::NumericMatrix& m, int64_t k, int64
   if(verbose) Rcpp::Rcout<<"creating space of type "<<indexType<<" done\n";
   n2::Hnsw index(m.ncol(), indexType);
   if(verbose) Rcpp::Rcout<<"adding data ... "<<std::flush;
-  for(int64_t i=0;i<m.nrow();i++) {
+  for(int i=0;i<m.nrow();i++) {
     Rcpp::NumericVector nv = m.row(i);
     std::vector<float> v(nv.begin(),nv.end());
     index.AddData(v);
@@ -54,9 +54,9 @@ Eigen::SparseMatrix<double> n2Knn(const Rcpp::NumericMatrix& m, int64_t k, int64
   index.Build(M, MaxM0, -1, nThreads);
   if(verbose) Rcpp::Rcout<<"done"<<std::endl;
 
-  int64_t ef_search = k*ef_search_multiplier;
+  int ef_search = k*ef_search_multiplier;
 
-  int64_t nanswers=k*m.nrow();
+  int nanswers=k*m.nrow();
 
   if(verbose) Rcpp::Rcout<<"querying ... "<<std::flush;
 
@@ -64,12 +64,12 @@ Eigen::SparseMatrix<double> n2Knn(const Rcpp::NumericMatrix& m, int64_t k, int64
   std::vector<T> tripletList;
   tripletList.reserve(nanswers);
 
-  for(int64_t i=0;i<m.nrow();i++) {
+  for(int i=0;i<m.nrow();i++) {
     std::vector<std::pair<int, float> > result;
     index.SearchById(i, k, ef_search, result);
-    int64_t nr=result.size();
+    int nr=result.size();
     if(nr>k) nr=k;
-    for(int64_t j=0;j<nr;j++) {
+    for(int j=0;j<nr;j++) {
       //int64_t l=i*k+j;
       //cout<<"i="<<i<<"; j="<<j<<"; n="<<result[j].first<<"; d="<<result[j].second<<endl;
       tripletList.push_back(T(result[j].first,i,result[j].second));
@@ -83,7 +83,7 @@ Eigen::SparseMatrix<double> n2Knn(const Rcpp::NumericMatrix& m, int64_t k, int64
 
 
 // [[Rcpp::export]]
-Eigen::SparseMatrix<double> n2CrossKnn(const Rcpp::NumericMatrix& mA, const Rcpp::NumericMatrix& mB, int64_t k, int64_t nThreads=10, bool verbose=true, std::string indexType="angular", int64_t M=12, int64_t MaxM0=24, float ef_search_multiplier=50, bool quiet=false) {
+Eigen::SparseMatrix<double> n2CrossKnn(const Rcpp::NumericMatrix& mA, const Rcpp::NumericMatrix& mB, int k, int nThreads=10, bool verbose=true, std::string indexType="angular", int M=12, int MaxM0=24, float ef_search_multiplier=50, bool quiet=false) {
 
   Eigen::SparseMatrix<double> mat(mB.nrow(),mA.nrow());
 
@@ -98,7 +98,7 @@ Eigen::SparseMatrix<double> n2CrossKnn(const Rcpp::NumericMatrix& mA, const Rcpp
   if(verbose) Rcpp::Rcout<<"creating space of type "<<indexType<<" done\n";
   n2::Hnsw index(mB.ncol(), indexType);
   if(verbose) Rcpp::Rcout<<"adding data ... "<<std::flush;
-  for(int64_t i=0;i<mB.nrow();i++) {
+  for(int i=0;i<mB.nrow();i++) {
     Rcpp::NumericVector nv = mB.row(i);
     std::vector<float> v(nv.begin(),nv.end());
     index.AddData(v);
@@ -109,22 +109,22 @@ Eigen::SparseMatrix<double> n2CrossKnn(const Rcpp::NumericMatrix& mA, const Rcpp
   index.Build(M, MaxM0, -1, nThreads);
   if(verbose) Rcpp::Rcout<<"done"<<std::endl;
 
-  int64_t ef_search = k*ef_search_multiplier;
+  int ef_search = k*ef_search_multiplier;
 
-  int64_t nanswers=k*mA.nrow();
+  int nanswers=k*mA.nrow();
   std::vector<T> tripletList;
   tripletList.reserve(nanswers);
 
   if(verbose) Rcpp::Rcout<<"querying ... "<<std::flush;
 
-  for(int64_t i=0;i<mA.nrow();i++) {
+  for(int i=0;i<mA.nrow();i++) {
     Rcpp::NumericVector nv = mA.row(i);
     std::vector<float> v(nv.begin(),nv.end());
     std::vector<std::pair<int, float> > result;
     index.SearchByVector(v, k, ef_search, result);
-    int64_t nr=result.size();
+    int nr=result.size();
     if(nr>k) nr=k;
-    for(int64_t j=0;j<nr;j++) {
+    for(int j=0;j<nr;j++) {
       //int64_t l=i*k+j;
       tripletList.push_back(T(result[j].first,i,result[j].second));
     }
